# doc/openapi.yaml
openapi: 3.1.0
info:
  title: Vectis HTTP API
  version: v1
  description: >
    Vectis protects data throughout its lifecycle. The API creates key material,
    validates keys, publishes public keys, signs message hashes, exchanges
    protected messages between Vectis instances, and encrypts/decrypts internal
    messages.
servers:
  - url: http://127.0.0.1:3000
    description: Local default server over HTTP
  - url: https://127.0.0.1:3000
    description: Local HTTPS server when VECTIS_MODE=prod
...
security: []
paths:
...
  /keys:
    get:
      tags: [Keys]
      summary: List keys loaded in local memory
      description: Public endpoint. Lists keys currently loaded in this node's in-memory state. Decrypted lifecycle properties are not exposed here.
      responses:
        "200":
          description: Loaded key list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListKeysOutput"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Keys]
      summary: Create operational key material
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeysInput"
            examples:
              profileOnly:
                value:
                  tag: payments
                  profile: hybrid-high-assurance-v1
              allowOverrides:
                value:
                  tag: dev-key
                  profile: hybrid-performance-v1
                  symmetric_algorithm: AES-256/GCM
      responses:
        "200":
          description: Created key KID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateKeysResponse"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        default:
          $ref: "#/components/responses/Error"
  /keys/reload:
    post:
      tags: [Keys]
      summary: Refresh local key state from storage
      description: Administrative operation. Reloads and decrypts keys plus encrypted lifecycle properties from storage into this node's memory. Uses POST because it changes server memory state.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Refreshed loaded key list with decrypted properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListKeysPropertiesOutput"
        default:
          $ref: "#/components/responses/Error"
  /keys/properties:
    get:
      tags: [Keys]
      summary: List keys loaded in local memory with properties
      description: Administrative endpoint. Lists keys and decrypted lifecycle properties currently loaded in this node's in-memory state. The info field is AAD for keys; properties_info is AAD for properties.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Loaded key list with properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListKeysPropertiesOutput"
        default:
          $ref: "#/components/responses/Error"
  /keys/properties/{kid}:
    get:
      tags: [Keys]
      summary: Return decrypted properties for one key
      description: Administrative endpoint. Returns decrypted lifecycle properties for one key. If the key is not currently loaded in memory, Vectis attempts to load it from storage first.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Kid"
      responses:
        "200":
          description: Key properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyPropertiesOutput"
        default:
          $ref: "#/components/responses/Error"
  /lifecycle/{kid}:
    post:
      tags: [Keys]
      summary: Update key lifecycle metadata
      description: Administrative endpoint. Updates encrypted lifecycle metadata for an operational key. Lifecycle status is enforced by cryptographic operations. Allowed transitions are active to disabled/retired/compromised/destroyed, and disabled to active. Retired, compromised, and destroyed are terminal states. Same-status transitions are rejected.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Kid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLifecycleInput"
            examples:
              disabled:
                value:
                  status: disabled
                  reason: maintenance window
      responses:
        "200":
          description: Updated lifecycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateLifecycleOutput"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        default:
          $ref: "#/components/responses/Error"
...
  /config/reload:
    post:
      tags: [Config]
      summary: Reload the unified signed config
      description: Administrative endpoint requiring root or an admin client. Reloads the unified config (routes, remote_routes, permissions, fpe_profiles, tokenization_profiles, mac_profiles, commitment_profiles, masking_profiles, sharing_profiles) only after VECTIS_CONFIG_PATH and VECTIS_CONFIG_SIGN_PATH are successfully read, verified, and validated. Every route and profile kid must exist in the keys currently loaded in memory. A missing config reloads to empty sections; an invalid signature, invalid file, or unloaded kid returns an error and keeps the previous in-memory config. If config.json changed but config_sign.json signs older content, reload keeps the previous state and returns a warning.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Reloaded unified config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReloadConfigResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Kid:
      name: kid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Kid"
  responses:
    Error:
      description: Public error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RequestBodyTooLarge:
      description: Request body exceeds the internal 2 MiB limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: request body exceeds maximum allowed size
  schemas:
    Kid:
      type: string
      pattern: "^[0-9a-fA-F]{64}$"
      example: f55f086e75b58ac4dfaffd3e75c90d25719281df90e87880145fb9f2e32f2eed
...
    TextField:
      type: string
      minLength: 1
      pattern: '^[^\x00-\x1F\x7F-\x9F]+$'
...
    AadSafeText:
      type: string
      minLength: 1
      pattern: '^[^\x00-\x1F\x7F-\x9F;=]+$'
      description: Text safe to embed in a key=value;... AAD string (no ';' or '=').
    Timestamp:
      type: string
      pattern: "^[0-9]+$"
      example: "1782058090"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/TextField"
          example: invalid request
...
    CreateKeysInput:
      type: object
      additionalProperties: false
      properties:
        tag:
          $ref: "#/components/schemas/AadSafeText"
          example: payments
        profile:
          $ref: "#/components/schemas/CryptoProfile"
        hash_algorithm:
          $ref: "#/components/schemas/HashAlgorithm"
        symmetric_algorithm:
          $ref: "#/components/schemas/SymmetricAlgorithm"
        eddsa_algorithm:
          $ref: "#/components/schemas/EdDsaAlgorithm"
        xecdh_algorithm:
          $ref: "#/components/schemas/XecdhAlgorithm"
        ml_dsa_variant:
          $ref: "#/components/schemas/MlDsaVariant"
        ml_kem_variant:
          $ref: "#/components/schemas/MlKemVariant"
      description: >
        When VECTIS_CRYPTO_POLICY=profile-only, only tag and profile are accepted.
        Individual algorithm fields are accepted only when VECTIS_CRYPTO_POLICY=allow-overrides.
        Optional fields may be omitted, but explicit null values are rejected.
    CreateKeysResponse:
      type: object
      required: [kid]
      properties:
        kid:
          $ref: "#/components/schemas/Kid"
    ListKeysOutput:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            type: object
            required: [kid, info]
            properties:
              kid:
                $ref: "#/components/schemas/Kid"
              info:
                $ref: "#/components/schemas/TextField"
                example: version=v1;hostname=localhost;type=ops-keys;cipher=AES-192/GCM;tag=payments;profile=hybrid-high-assurance-v1;timestamp=1782058090
    ListKeysPropertiesOutput:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            type: object
            required: [kid, info, properties_info, properties]
            properties:
              kid:
                $ref: "#/components/schemas/Kid"
              info:
                $ref: "#/components/schemas/TextField"
              properties_info:
                $ref: "#/components/schemas/TextField"
              properties:
                $ref: "#/components/schemas/OpsKeyProperties"
    KeyPropertiesOutput:
      type: object
      required: [kid, info, properties_info, properties]
      properties:
        kid:
          $ref: "#/components/schemas/Kid"
        info:
          $ref: "#/components/schemas/TextField"
        properties_info:
          $ref: "#/components/schemas/TextField"
        properties:
          $ref: "#/components/schemas/OpsKeyProperties"
    OpsKeyProperties:
      type: object
      required: [version, profile, tag, created_at, lifecycle]
      properties:
        version:
          const: 1
        profile:
          enum:
            - hybrid-performance-v1
            - hybrid-standard-v1
            - hybrid-high-assurance-v1
            - hybrid-long-term-v1
            - custom
        tag:
          $ref: "#/components/schemas/TextField"
          example: payments-prod
        created_at:
          $ref: "#/components/schemas/Timestamp"
        lifecycle:
          $ref: "#/components/schemas/OpsKeyLifecycle"
    OpsKeyLifecycle:
      type: object
      required: [status, reason, changed_at]
      properties:
        status:
          $ref: "#/components/schemas/LifecycleStatus"
        reason:
          $ref: "#/components/schemas/TextField"
          maxLength: 128
          example: initial creation
        changed_at:
          type: string
          example: "1782058090"
    LifecycleStatus:
      type: string
      enum: [active, disabled, retired, compromised, destroyed]
    UpdateLifecycleInput:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status:
          $ref: "#/components/schemas/LifecycleStatus"
        reason:
          $ref: "#/components/schemas/TextField"
          maxLength: 128
          example: maintenance window
    UpdateLifecycleOutput:
      type: object
      required: [kid, lifecycle]
      properties:
        kid:
          $ref: "#/components/schemas/Kid"
        lifecycle:
          $ref: "#/components/schemas/OpsKeyLifecycle"
...
    ReloadConfigResponse:
      type: object
      required: [status, routes_loaded, remote_routes_loaded, clients_loaded, fpe_profiles_loaded, tokenization_profiles_loaded, mac_profiles_loaded, commitment_profiles_loaded, masking_profiles_loaded, sharing_profiles_loaded]
      properties:
        status:
          const: reloaded
        warning:
          type: string
          description: Present only when config.json has changes not covered by config_sign.json and the previous in-memory config was kept.
        routes_loaded:
          type: integer
          minimum: 0
        remote_routes_loaded:
          type: integer
          minimum: 0
        clients_loaded:
          type: integer
          minimum: 0
        fpe_profiles_loaded:
          type: integer
          minimum: 0
        tokenization_profiles_loaded:
          type: integer
          minimum: 0
        mac_profiles_loaded:
          type: integer
          minimum: 0
        commitment_profiles_loaded:
          type: integer
          minimum: 0
        masking_profiles_loaded:
          type: integer
          minimum: 0
        sharing_profiles_loaded:
          type: integer
          minimum: 0
...
    CryptoProfile:
      enum:
        - hybrid-performance-v1
        - hybrid-standard-v1
        - hybrid-high-assurance-v1
        - hybrid-long-term-v1
    HashAlgorithm:
      enum:
        - BLAKE2b(160)
        - BLAKE2b(224)
        - BLAKE2b(256)
        - BLAKE2b(384)
        - BLAKE2b(512)
        - SHA-224
        - SHA-256
        - SHA-384
        - SHA-512
        - SHA-512-256
        - SHA-3(224)
        - SHA-3(256)
        - SHA-3(384)
        - SHA-3(512)
        - Whirlpool
    SymmetricAlgorithm:
      enum:
        - ChaCha20Poly1305
        - AES-128/GCM
        - AES-192/GCM
        - AES-256/GCM
    EdDsaAlgorithm:
      enum: [Ed25519, Ed448]
    XecdhAlgorithm:
      enum: [X25519, X448]
    MlDsaVariant:
      enum: [ML-DSA-44, ML-DSA-65, ML-DSA-87]
    MlKemVariant:
      enum: [ML-KEM-512, ML-KEM-768, ML-KEM-1024]